requirements.txt
streamlit
bcrypt
Pillow
openai
requests
python-dotenv
